fix(display): standardize rotation handling and initialization order
Fix inconsistent display rotation behavior where configuration (e.g., DISPLAY_ROTATION=2) was not correctly applied. Standardize rotation input to handle both index (0-3) and degree values (90-270) across all drivers and ensure state is propagated before initialization.

- Standardize rotation input in TFT_eSPI_Drawer and SDL2_Drawer using uint16_t
- Fix initialization order in Renderer to apply rotation before driver init
- Update main.cpp and main_native.cpp to use new PHYSICAL_DISPLAY_* macros
- Remove obsolete Config.h dependency from main.cpp
- Add DISPLAY_ROTATION to native environment in platformio.ini

+++ include/drivers/esp32/TFT_eSPI_Drawer.h
@@ -35,7 +35,7 @@ public:
      * @brief Sets the screen rotation.
      * @param rotation 0-3 corresponding to 0, 90, 180, 270 degrees.
      */
-    void setRotation(uint8_t rotation) override;
+    void setRotation(uint16_t rotation) override;
 
     /**
      * @brief Fills the sprite buffer with black color.
