+++ src/drivers/native/SDL2_Drawer.cpp
@@ -74,7 +74,7 @@ void pr32::drivers::native::SDL2_Drawer::init() {
     memset(pixels, 0, logicalWidth * logicalHeight * sizeof(uint16_t));
 }
 
-void pr32::drivers::native::SDL2_Drawer::setRotation(uint8_t rot) {
+void pr32::drivers::native::SDL2_Drawer::setRotation(uint16_t rot) {
     // Standardize rotation to index 0-3 (0, 90, 180, 270)
     if (rot == 90) rotation = 1;
     else if (rot == 180) rotation = 2;
